#27 Rename submodule (api → core)
