+++ xarrayspecs/__init__.py
@@ -1,7 +1,7 @@
 __all__ = [
     # submodules
-    "api",
     "convert",
+    "core",
     "typing",
     # aliases
     "AsDataArray",
@@ -21,5 +21,5 @@ __all__ = [
 __version__ = "0.6.0"
 
 # dependencies
-from . import api, convert, typing
-from .api import *
+from . import convert, core, typing
+from .core import *
